requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
paho-mqtt>=1.6.1
numpy>=1.24.0
scipy>=1.11.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
langgraph>=0.0.20
langchain-core>=0.1.0
openai>=1.3.0
onnxruntime>=1.16.0
redis>=5.0.0
python-dotenv>=1.0.0
